refactored connect.js
